gdb: Change how frames are selected for 'frame' and 'info frame'.
The 'frame' command, and thanks to code reuse the 'info frame' and
'select-frame' commands, currently have an overloaded mechanism for
selecting a frame.

These commands take one or two parameters, if it's one parameter then
we first try to use the parameter as an integer to select a frame by
number.  If that fails then we treat the parameter as an address and
try to select a stack frame by stack-address.  If we still have not
selected a stack frame, or we initially had two parameters, then GDB
allows the user to view a stack frame that is not part of the current
backtrace.  Internally, a new frame is created with the given stack
and pc addresses, and this is shown to the user.

The result of this is that a typo by the user, entering the wrong stack
frame number for example, can result in a brand new frame being viewed
rather than an error.

The purpose of this commit is to remove this overloading, while still
offering the same functionality through some new sub-commands.  By
making the default behaviour of 'frame' (and friends) be to select a
stack frame by number, it is hoped that enough backwards-compatibility
is maintained that users will not be overly inconvenienced.

The 'frame', 'select-frame', and 'info frame' commands now all take a
frame specification string as an argument, this string can be any of the
following:

  (1) An integer.  This is treated as a frame number.  If a frame with
  that number does not exist then the user gets an error.

  (2) A string like 'number <NUMBER>', where <NUMBER> is the frame to
  select as in option (1) above.

  (3) A string like 'address <STACK-ADDRESS>', where <STACK-ADDRESS>
  is a stack-frame address.  If there is no frame for this address
  then the user gets an error.

  (4) A string like 'function <NAME>', where <NAME> is a function name,
  the inner most frame for function <NAME> is selected.  If there is no
  frame for function <NAME> then the user gets an error.

  (5) A string like 'view <STACK-ADDRESS>', this views a new frame
  with stack address <STACK-ADDRESS>.

  (6) A string like 'view <STACK-ADDRESS> <PC-ADDRESS>', this views
  a new frame with stack address <STACK-ADDRESS> an the pc <PC-ADDRESS>.

This change assumes that the most common use of the commands like
'frame' is to select a frame by number, it is for this reason that
this is the behaviour that is kept for backwards compatibility.  Any
of the alternative behaviours, which are assumed to be less used, now
require a change in user behaviour.

The MI command '-stack-select-frame' has also changed in the same way.
The default behaviour is to select a frame by number, but the other
selection methods are also available.

gdb/ChangeLog:

	(NEWS): Mention changes to frame related commands.
	* cli/cli-decode.c (add_cmd_suppress_notification): New function.
	(add_prefix_cmd_suppress_notification): New function.
	(add_com_suppress_notification): Call
	add_cmd_suppress_notification.
	* command.h (add_cmd_suppress_notification): Declare.
	(add_prefix_cmd_suppress_notification): Declare.
	* mi/mi-cmd-stack.c (mi_cmd_stack_select_frame): Use new function
	select_frame_from_spec.
	* stack.c (find_frame_for_function): Add declaration.
	(find_frame_for_address): New function.
	(parse_frame_specification): Delete function.
	(frame_selection_by_function_completer): New function.
	(info_frame_command): Rename to...
	(info_frame_command_core): ...this, and update parameter types.
	(select_frame_command): Rename to...
	(select_frame_command_core): ...this, and update parameter types.
	(frame_command): Rename to...
	(frame_command_core): ...this, and update parameter types.
	(class frame_command_helper): New class to wrap implementations of
	frame related sub-commands.
	(select_frame_from_spec): New function.
	(frame_apply_cmd_list): New static global.
	(frame_cmd_list): Make static.
	(select_frame_cmd_list): New global for sub-commands.
	(info_frame_cmd_list): New global for sub-commands.
	(_initialize_stack): Register sub-commands for 'frame',
	'select-frame', and 'info frame'.  Update 'frame apply' commands
	to use frame_apply_cmd_list.  Move function local static
	frame_apply_list to file static frame_apply_cmd_list for
	consistency.
	* stack.h (select_frame_command): Delete declarationn.
	(select_frame_from_spec): Declare new function.

gdb/doc/ChangeLog:

	* gdb.texinfo (Selection): Rewrite documentation for 'frame' and
	'select-frame' commands.
	(Frame Info): Rewrite documentation for 'info frame' command.
	(GDB/MI Stack Manipulation): Update description of
	'-stack-select-frame'.

gdb/testsuite/ChangeLog:

	* gdb.base/frame-selection.exp: New file.
	* gdb.base/frame-selection.c: New file.
	* gdb.mi/mi-frame-selection.exp: New file.
	* gdb.mi/mi-frame-selection.c: New file.