cpu/opcodes: Sync up fr30.cpu and generated opcodes files
In commit b6518b3 (Fix compile time warnings generated when compiling
with clang.) the generated fr30-ibld.c file was modified, but the
underlying fr30.cpu file, from which it is generated was not.

This commit updates the fr30.cpu file in such a way that the C code
generated in fr30-ibld.c avoids the undefined behaviour that b6518b3 was
trying to avoid, while also being something that CGEN can handle without
any changes.

cpu/ChangeLog:

	* fr30.cpu (f-m4): Adjust extract handler to avoid shifting a
	negative value (undefined behaviour in C).

opcodes/ChangeLog:

	* fr30-ibld.c: Regenerate from latest cpu/fr30.cpu file.