gdb: Fix examine-backward testcase
This test has a number of issues, first, it expect a zero address to
have some hex characters followed by '00', this is only true for targets
that, even for a zero address, have some non-zero bits in the
most-significant parts of the address.  For targets that don't have this
characteristic, then the printed address will be just '0x0'.

Second, the test assumes that if you can read from address zero, then
you can read from all memory, this is not true for all targets.