Mostly extending the dummy call mechanism
Massive improvements to the dummy call mechanism.  Now shares code with
the return value read/write code, and also does a better job following
some of the struct flattening rules that are in the ABI.

Small teak to some of the 'info registers' handling code too.

New test covers some of the structs with a single complex float case,
but I wonder if these would be better folded into the
gdb.base/structs.exp test script.