sim: Don't overwrite stored errno in sim_syscall_multi
The host syscall callback mechanism should take care of updating the
errcode within the CB_SYSCALL struct, and we should not be adjusting
the error code once the syscall has completed.  We especially, should
not be rewriting the syscall errcode based on the value of errno some
time after running the host syscall, as there is no guarantee that
errno has not be overwritten.

Lets consider why reading errno in sim_syscall_multi is a bad idea.
To perform a syscall we call cb_syscall (in syscall.c), there are two
exit paths from this function, these are called FinishSyscall and
ErrorFinish.

In FinishSyscall we store the syscall result in 'sc->result', and the
error code is transated to target encoding, and stored in
'sc->errcode'.

In ErrorFinish, we again store the syscall result in 'sc->result', and
fill in 'sc->errcode' by fetching the actual errno from the host with
the 'cb->get_errno' callback.

In both cases 'sc->errcode' should have been filled in with an
appropriate value.

However, lets consider a specific syscall CB_SYS_open, in this case,
the first thing we do is fetch the path to open from the target with
'get_path', if this fails then the errcode is returned, and we call
FinishSyscall.  Notice that in this case, no host syscall may have
been performed, and so reading errno makes absolutely no sense.

This commit removes from sim_syscall_multi the rewriting of
sc->errcode based on the value of errno.

sim/common/ChangeLog:

	* sim-syscall.c (sim_syscall_multi): Don't update sc->errcode at
	this point, it should have already been set in cb_syscall.