gdb/riscv: Improved register names for CSRs
Only return register names for the actual CSR register numbers.  Use
the opcode/riscv-opc.h file to generate the register name returning
code, in order to correspond to the register numbers which are
generated using the same file.  Any other register number should
return NULL to indicate it is not a valid register.

gdb/ChangeLog:

	* riscv-tdep.c (riscv_register_name): Only return names for actual
	CSR register numbers, return NULL for anything else.