gdb/riscv: remove extra caching of misa register
The RiscV had a mechanism in place to cache the contents of the misa
register per-inferior, the original intention behind this was to
reduce the number of times the misa register had to be read, but it
was pointed out on the mailing list that the register cache will mean
the register is only accessed once each time GDB stops.

As such, until it can be shown that there's a real need for additional
caching, this commit removes all of the additional caching, and just
accesses the misa register as any other register.

gdb/ChangeLog:

	* riscv-tdep.c (struct riscv_inferior_data): Delete.
	(riscv_read_misa_reg): Don't cache value read into inferior data.
	(riscv_new_inferior_data): Delete.
	(riscv_inferior_data_cleanup): Delete.
	(riscv_inferior_data): Delete.
	(riscv_invalidate_inferior_data): Delete.
	(_initialize_riscv_tdep): Remove initialisation of inferior data.